+++ BLCMM/src/CHANGELOG.md
@@ -20,6 +20,9 @@ BLCMM Changelog
 - Improved Steam Library Folder detection using StidOfficial's SteamVDF
 - Moved "offline" checkbox to new "Dangerous settings" dialog.  Mods will be
   saved in offline mode by default, and it's recommended to keep it that way.
+- "Restore defaults" button on settings screen will only restore the settings
+  shown on the active screen (this prevents clearing the recently-viewed file
+  list, for instance)
 
 (changelog entries to confirm, etc -- from original 1.2.1 plans ages ago)
 - When creating and saving a new file, it now opens on the next boot